+++ resources/js/field.js
@@ -1,22 +1,15 @@
-Nova.booting((Vue, router, store) => {
-    // Vue.config.devtools = true;
+Nova.booting(Vue => {
+    Vue.config.devtools = true;
 
-    Vue.filter('striphtml', function (value) {
-        var div = document.createElement("div");
-        div.innerHTML = value;
-        var text = div.textContent || div.innerText || "";
-        return text;
-    });
+    Vue.component('index-nova-translatable', require('./components/IndexField.vue'));
+    Vue.component('detail-nova-translatable', require('./components/DetailField.vue'));
+    Vue.component('form-nova-translatable', require('./components/FormField.vue'));
 
-    Vue.component('index-nova-translatable', require('./components/IndexField'))
-    Vue.component('detail-nova-translatable', require('./components/DetailField'))
-    Vue.component('form-nova-translatable', require('./components/FormField'))
+    Vue.component('index-nova-translatable-slug-field', require('./components/Slug/IndexField.vue'));
+    Vue.component('detail-nova-translatable-slug-field', require('./components/Slug/DetailField.vue'));
+    Vue.component('form-nova-translatable-slug-field', require('./components/Slug/FormField.vue'));
 
-    Vue.component('index-nova-translatable-slug-field', require('./components/Slug/IndexField'))
-    Vue.component('detail-nova-translatable-slug-field', require('./components/Slug/DetailField'))
-    Vue.component('form-nova-translatable-slug-field', require('./components/Slug/FormField'))
-
-    Vue.component('index-nova-translatable-code-field', require('./components/Code/IndexField'))
-    Vue.component('detail-nova-translatable-code-field', require('./components/Code/DetailField'))
-    Vue.component('form-nova-translatable-code-field', require('./components/Code/FormField'))
-})
+    Vue.component('index-nova-translatable-code-field', require('./components/Code/IndexField.vue'));
+    Vue.component('detail-nova-translatable-code-field', require('./components/Code/DetailField.vue'));
+    Vue.component('form-nova-translatable-code-field', require('./components/Code/FormField.vue'));
+});