+++ src/Http/Middleware/Authorize.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rslanzi\NovaTranslatable\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
+
+class Authorize
+{
+    public function handle(Request $request, Closure $next): Response
+    {
+        return $next($request);
+    }
+}