+++ webpack.mix.js
@@ -1,6 +1,6 @@
 let mix = require('laravel-mix')
 
 mix
-  .setPublicPath('dist')
-  .js('resources/js/field.js', 'js')
-  .sass('resources/sass/field.scss', 'css')
+    .setPublicPath('dist')
+    .sass('resources/sass/field.scss', 'css')
+    .js('resources/js/field.js', 'js');