+++ CHANGELOG.md
@@ -4,8 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.0] - 2021-05-09
+### Breaking Changes
+- Upgraded to Laravel Mix 6
+- Renamed Service Provider
+
+### Changes
 - Refactoring
+- Added customization options to Slug Field
+
+### Added
 - Tests
 
 ## [0.2.3] - 2021-03-31